+++ lib/just_motion.dart
@@ -1,15 +1,16 @@
 library just_motion;
 
+import 'dart:async';
 import 'dart:ui';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter/scheduler.dart';
-import 'dart:async';
 
-part 'src/motion_ticker.dart';
 part 'src/ease.dart';
-part 'src/motion_value.dart';
 part 'src/motion_extensions.dart';
-part 'src/spring.dart';
+part 'src/motion_ticker.dart';
+part 'src/motion_value.dart';
 part 'src/motion_widget.dart';
+part 'src/spring.dart';