+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.6+22
+- renamed `MotionState` to `MotionStatus`, to keep it idiomatic with Flutter's `AnimationController`.
+- renamed `motion.state` to `motion.status`.
+- added `MotionState.disposed` notification.
+- fix notifications for `MotionState.moving`
+- improved `toString()` on `Ease` and `Spring` motion variants, to describe better the motion instance.
+- removed the "microtask delay" for new states notifications with `statusListener`.
+- added missing hot-reload support code.
+- EXPERIMENTAL: added instance auto de-registration for `MotionValue` instances created inside `build(BuildContext)`, this is
+  still experimental, and still have to be tested in multiple use-cases of motion, to see if it brings collateral issues. 
+- add some code docs.
+
 ## 0.0.5+15
 - `MotionValue` forced to dispose on `reassembling` (hot-reload).
 - add some code docs.